double fields, names, Switch class

+++ app/src/main/java/tuehomework/thermostat/Thermostat/OnTemperatureChangedListener.java
@@ -4,5 +4,5 @@ package tuehomework.thermostat.Thermostat;
  * Created by Platon on 28.05.2015.
  */
 public interface OnTemperatureChangedListener {
-    void onTemperatureChanged(int new_temp);
+    void onTemperatureChanged(double new_temp);
 }
